fix: address issue #401

+++ includes/class-cache.php
@@ -804,8 +804,13 @@ if ( ! class_exists( 'PerformanceOptimise\Inc\Cache' ) ) {
 		 * @since 1.1.1
 		 */
 		public static function clear_cache( $url_path = null ): bool {
-			delete_transient( 'wppo_cache_size' );
-			delete_transient( 'wppo_total_js_css' );
+			if ( function_exists( 'wp_cache_get_salted' ) ) {
+				$salt = (int) get_option( 'wppo_cache_last_cleared', 0 ) + 1;
+				update_option( 'wppo_cache_last_cleared', $salt, false );
+			} else {
+				delete_transient( 'wppo_cache_size' );
+				delete_transient( 'wppo_total_js_css' );
+			}
 
 			$instance = new self();
